+++ lib/dart_date.dart
@@ -1,3 +1,0 @@
-library dart_date;
-
-export 'src/dart_date.dart';